【优化】sample文件统一处理 【修改】MSH命令变更
