+++ class/air720/ppp_device_air720.h
@@ -13,13 +13,6 @@
 
 #include <ppp_device.h>
 
-/* Air720  ppp_device base from ppp_device */
-struct ppp_air720
-{
-    struct ppp_device  device;          /* ppp_device struct in ppp_air720 */
-    enum ppp_trans_type type;           /* the type is used to establish a ppp connection */
-};
-
-extern int ppp_air720_register(struct ppp_air720 *air720, const char *dev_name, const char *rely_name, void *user_data);
+extern int ppp_air720_register(struct ppp_sample *air720, const char *dev_name, const char *rely_name, void *user_data);
 
 #endif  /* __PPP_AIR720_H__ */
